Fix 1: In case of fragmentation with VLANs, there was a bug in the datapath where after inserting the VLAN header in the packet mbuf structure, the pkt datastructure was not updated using pkt_push() call. This was causing corruption to the packet header when fragmentation was happening in dpdk_fragment_packet(). Fixed the same.
Fix 2:
In case of bonding, the rx error counter was incorrectly being incremented even though
packets were correctly received by the NIC. The XEC error counter in the NIC was
incrementing due to 0 UDP checksum. In such cases, we need to subtract the XEC count
to get the acutal error count. This was not happening in bonded case. Fixed the same.

Closes-bug: #1532319

+++ dpdk/vr_dpdk_interface.c
@@ -1309,6 +1309,10 @@ dpdk_port_stats_update(struct vr_interface *vif, unsigned lcore_id)
     }
 }
 
+/* For non-bonded interfaces, use the xstats from the rte_eth_xstats_get() API.
+ * For bonded interfaces, this API is not available, so instead use xstats from
+ * it's individual slave interfaces
+ */
 static void 
 vr_dpdk_eth_xstats_get(uint32_t port_id, struct rte_eth_stats *eth_stats)
 {
@@ -1350,7 +1354,7 @@ vr_dpdk_eth_xstats_get(uint32_t port_id, struct rte_eth_stats *eth_stats)
     } while (port_num < ethdev->ethdev_nb_slaves);
 
     /* Stats cannot go negative */
-    if (eth_stats->ierrors < 0)
+    if ((int64_t)eth_stats->ierrors < 0)
         eth_stats->ierrors = 0;
 }
